requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
altgraph==0.17.4
async-timeout==4.0.2
attrs==23.1.0
cchardet==2.1.7
charset-normalizer==3.1.0
discord-protos==0.0.2
discord.py-self==2.0.0
frozenlist==1.3.3
idna==3.4
importlib_metadata==8.5.0
macholib==1.16.3
multidict==6.0.4
packaging==24.1
protobuf==4.23.2
pyinstaller==6.10.0
pyinstaller-hooks-contrib==2024.8
yarl==1.9.2
zipp==3.20.2
